Separate cpp_dfm()
